Rename scrapy/statscol.py to scrapy/statscollectors.py
