+++ scrapy/statscollectors.py
