+++ tests/test_spidermiddleware_depth.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 from scrapy.spidermiddlewares.depth import DepthMiddleware
 from scrapy.http import Response, Request
 from scrapy.spider import Spider
-from scrapy.statscol import StatsCollector
+from scrapy.statscollectors import StatsCollector
 from scrapy.utils.test import get_crawler
